add the fee payment app.
make test will issue a call to confirm that the datacapture app.
can call the fee payment app's API

The fee payment app. needs to know its own FQDN, to return the
correct URL to which the user should be redirected. Currently,
the URL returned will not work, because it is 'http://fees:3000/...'
when it should be 'http://localhost:4000/...'