consolidate to one docker-compose environment
There were two - one running everything in
'production' mode and one running in 'dev' mode
with source code volume mounted into containers.
In practice, I think we're only ever going to want
to use this to run the system in dev mode, so the
extra overhead of maintaining multiple versions
is unnecessary.