Put back automatic repository updating
In the first cut of the script, it would update the git repos if they
already existed.  I took that out for brevity.  On reflection, it seemed
reasonable to put it back in as the likely target audience for regular
use of the script is unlikely to have large amounts of command-line
experience; the fewer commands, the better.